fixed inconsistent background calculation

+++ data/src/PeakFinder.cpp
@@ -68,7 +68,8 @@ bool PeakFinder::find(std::vector<std::shared_ptr<IData>> numors)
             //progressDialog->setLabelText("Computing background level...");
             //progressDialog->show();
 
-            _median = numor->getBackgroundLevel(_handler)+1;
+            // jmf: why do we round median to an integer??
+            _median = static_cast<int>(numor->getBackgroundLevel(_handler))+1;
 
             //progressDialog->close();
         }
